Added 403 errors for views
Enhancements made to Destroy and Update views: now, attempting to access others' data triggers a 403 error. Previously, the queryset contained only habits belonging to the current user, and for Retrieve views, it included public records as well. With the addition of permission_classes and queryset modifications, stricter protection is enforced, resulting in a 403 error upon attempting to access other users' data.